Removed name of area, it was never used.
