+++ kiwi.vcxproj
@@ -163,7 +163,6 @@
     <ClCompile Include="kiwi.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="area.h" />
     <ClInclude Include="city.h" />
     <ClInclude Include="config.h" />
     <ClInclude Include="matrix.h" />
